[REFACTOR] 함수 파일 위치 변경... 잘못썼다 ㅠ
